+++ PageRank/VectorMapMatrix.cpp
@@ -8,7 +8,7 @@ VectorMapMatrix::VectorMapMatrix() {
     width = 0;
 }
 
-VectorMapMatrix::VectorMapMatrix(uint h, uint w) : m(h, map<uint,float>()) , width(w) {}
+VectorMapMatrix::VectorMapMatrix(uint h, uint w) : m(h, map<uint,double>()) , width(w) {}
 
 size_t VectorMapMatrix::cantFilas() const {
     return m.size();
@@ -18,7 +18,7 @@ size_t VectorMapMatrix::cantColumnas() const {
     return width;
 }
 
-void VectorMapMatrix::asignar(uint f, uint c, const float value) {
+void VectorMapMatrix::asignar(uint f, uint c, const double value) {
     if (value==0) {
         m[f].erase(c);
     } else if (f < m.size() and c < width) {
@@ -26,8 +26,8 @@ void VectorMapMatrix::asignar(uint f, uint c, const float value) {
     }
 }
 
-const float VectorMapMatrix::at(uint f, uint c) const {
-    float result;
+const double VectorMapMatrix::at(uint f, uint c) const {
+    double result;
     if(m[f].count(c) == 0) {
         result = 0;
     } else {
@@ -36,7 +36,7 @@ const float VectorMapMatrix::at(uint f, uint c) const {
     return result;
 }
 
-float& VectorMapMatrix::operator[](pair<uint, uint> p) {
+double& VectorMapMatrix::operator[](pair<uint, uint> p) {
     if (p.first < m.size() and p.second < width) {
         return m[p.first][p.second];
     }
@@ -45,8 +45,8 @@ float& VectorMapMatrix::operator[](pair<uint, uint> p) {
 VectorMapMatrix VectorMapMatrix::operator+(VectorMapMatrix const &B) {
     if(cantFilas() == B.cantFilas() and cantColumnas() == B.cantColumnas()) {
         VectorMapMatrix result(cantFilas(), cantColumnas());
-        map<uint, float>::const_iterator it1 = m[0].begin();
-        map<uint, float>::const_iterator it2 = B.m[0].begin();
+        map<uint, double>::const_iterator it1 = m[0].begin();
+        map<uint, double>::const_iterator it2 = B.m[0].begin();
         uint f = 0;
         while (f < cantFilas()) {
             while(it1 != m[f].end() or it2 != B.m[f].end()) {
@@ -76,12 +76,12 @@ VectorMapMatrix VectorMapMatrix::operator+(VectorMapMatrix const &B) {
 VectorMapMatrix VectorMapMatrix::operator*(const VectorMapMatrix &B) {
     if(cantColumnas() == B.cantFilas()) {
         VectorMapMatrix result(cantFilas(), B.cantColumnas());
-        vector<vector<float> > sumasParciales(cantColumnas(),vector<float>(B.cantColumnas(), 0));
+        vector<vector<double> > sumasParciales(cantColumnas(),vector<double>(B.cantColumnas(), 0));
         uint f = 0; //Recorro sobre A, y cuando estoy en el elemento a(ij) multiplico por la fila j de B, sumándolo en result.
         while(f < cantFilas()) {
-            map<uint, float>::const_iterator it1 = m[f].begin();
+            map<uint, double>::const_iterator it1 = m[f].begin();
             while (it1 != m[f].end()) {
-                map<uint, float>::const_iterator it2 = B.m[it1->first].begin(); //Si el it1 esta en columna j, creo it2 en fila j.
+                map<uint, double>::const_iterator it2 = B.m[it1->first].begin(); //Si el it1 esta en columna j, creo it2 en fila j.
                 while(it2 != B.m[it1->first].end()) {
                     sumasParciales[f][it2->first] += it1->second * it2->second;//asigno en
                     // la fila f de A, la columna marcada por el iterador it2 de B, en result, lo que había mas el producto.
@@ -94,7 +94,7 @@ VectorMapMatrix VectorMapMatrix::operator*(const VectorMapMatrix &B) {
         f = 0;
         uint c = 0;
         while(f < sumasParciales.size()){
-            map<uint, float>::iterator hint_iterator = result.m[f].end();
+            map<uint, double>::iterator hint_iterator = result.m[f].end();
             while(c < sumasParciales[f].size()){
                 if(sumasParciales[f][c] != 0) {
                     result.m[f].insert(hint_iterator,make_pair(c, sumasParciales[f][c])); //como el elemento a insertar debe ir al final,
@@ -117,7 +117,7 @@ VectorMapMatrix VectorMapMatrix::mult(VectorMapMatrix const &B){
         for(uint i = 0; i < cantFilas(); ++i) { //Recorro las filas de A.
             for (auto it1 = m[i].cbegin(); it1 != m[i].cend(); ++it1) { //Recorro la i-ésima fila de A.
                 for(auto it2 = B.m[it1->first].begin(); it2 != B.m[it1->first].end(); ++it2) {  //Si el it1 esta en columna j, recorro la fila j de B.
-                    float sumando = it1->second * it2->second;  //"sumando" es a_ik*b_kj donde i = "i", k = "it1->first" y j = "it2->first".
+                    double sumando = it1->second * it2->second;  //"sumando" es a_ik*b_kj donde i = "i", k = "it1->first" y j = "it2->first".
                     auto it_res = result.m[i].find(it2->first); //Me fijo que hay en res_ij (= a_i1*b_1j + ... + a_i(k-1)*b_(k-1)j )
                     if(it_res == result.m[i].end()) //Si no estaba definido (era 0)...
                         result.m[i][it2->first] = sumando;  //lo defino como a_ik*b_kj.
@@ -136,10 +136,10 @@ VectorMapMatrix VectorMapMatrix::mult(VectorMapMatrix const &B){
     }
 }
 
-void VectorMapMatrix::operator*(float valor) {
-    float acum = 0;
+void VectorMapMatrix::operator*(double valor) {
+    double acum = 0;
     for (unsigned int f = 0; f<cantFilas(); f++) {
-        map<uint, float>::iterator row_iterator = m[f].begin();
+        map<uint, double>::iterator row_iterator = m[f].begin();
         while (row_iterator != m[f].end()) {
             if (valor != 0) {
                 row_iterator->second *= valor;
@@ -170,11 +170,11 @@ VectorMapMatrix VectorMapMatrix::permutar(unsigned int j, unsigned int i){
 
 }
 
-pair<vector<float>,short> VectorMapMatrix::EG(const VectorMapMatrix &mat, vector<float> bb) {
+pair<vector<double>,short> VectorMapMatrix::EG(const VectorMapMatrix &mat, vector<double> bb) {
 	unsigned int i,j;
-	vector<float> res(width,0);
+	vector<double> res(width,0);
 	short status = 0; //status default, el sistema tiene una unica solucion posible
-	float A_kk, A_jk;
+	double A_kk, A_jk;
 	VectorMapMatrix copy = VectorMapMatrix(mat);
 	VectorMapMatrix Mk = VectorMapMatrix(cantFilas(),width);
 	bool cont = false; //el bool da es false si en toda la columna de i en adelante es 0, es decir me tengo que saltear este paso
@@ -187,7 +187,7 @@ pair<vector<float>,short> VectorMapMatrix::EG(const VectorMapMatrix &mat, vector
 				cont = true;
 				if(copy.at(i,i) == 0){
 					copy[i].swap(copy[j]); //cambio de lugar las filas porque habia un 0 en la diagonal pero no en el resto de la columna
-                    float temp = bb[i];
+                    double temp = bb[i];
                     bb[i] = bb[j];         //como se cambiaron de lugar las filas, también se cambian de lugar los valores de "bb"
                     bb[j] = temp;
                 }
@@ -235,21 +235,21 @@ pair<vector<float>,short> VectorMapMatrix::EG(const VectorMapMatrix &mat, vector
 	return make_pair(res,status);
 }
 
-pair<vector<float>,short> VectorMapMatrix::EGPP(vector<float> bb) {
+pair<vector<double>,short> VectorMapMatrix::EGPP(vector<double> bb) {
     unsigned int i,j;
-    vector<float> res(width);
+    vector<double> res(width);
     short status = 0; //status default, el sistema tiene una unica solucion posible
     VectorMapMatrix copy = VectorMapMatrix(*this);
     for(i = 0; i < copy.cantFilas()-1; i++){ //itero sobre las filas, excepto la ultima porque ahi no tengo que hacer nada
-        float maximo_abs = copy.at(i,i);    //en cada iteración es igual al número más grande (en valor absoluto) del resto de la columna.
+        double maximo_abs = copy.at(i,i);    //en cada iteración es igual al número más grande (en valor absoluto) del resto de la columna.
         unsigned int fila_del_maximo = i;
-        list<pair<uint, float> > filas_a_ser_restadas;
+        list<pair<uint, double> > filas_a_ser_restadas;
         auto iter_de_la_lista_apuntando_al_maximo = filas_a_ser_restadas.begin();
         for(j = i+1; j < copy.cantFilas(); ++j){ //itero sobre las filas desde i+1 en adelante, para encontrar el valor de "maximo_abs".
             // De paso me guardo las filas en las que encuentro un valor != 0 en la columna "i" y me guardo dicho valor.
-            map<unsigned int, float>::const_iterator iter = copy[j].find(i);
+            map<unsigned int, double>::const_iterator iter = copy[j].find(i);
             if (iter != copy[j].cend()){    //Si el valor está definido, ie, es != 0:
-                const float& a_ji = iter->second;        //lo llamo a_ji y...
+                const double& a_ji = iter->second;        //lo llamo a_ji y...
                 filas_a_ser_restadas.emplace_back(j, a_ji);  //guardo la fila en la que está y su valor.
                 if(fabs(a_ji) > fabs(maximo_abs)){ //busco el máximo en valor absoluto
                     maximo_abs = a_ji;
@@ -264,16 +264,16 @@ pair<vector<float>,short> VectorMapMatrix::EGPP(vector<float> bb) {
             filas_a_ser_restadas.erase(iter_de_la_lista_apuntando_al_maximo);//También debo sacar de la lista de filas a restar, aquella que contiene al máximo y...
             if (filas_a_ser_restadas.begin()->first == i)    //si la fila "i" está en la lista...
                 filas_a_ser_restadas.begin()->first = fila_del_maximo;    //debe actualizarse el índice de su fila (pues la swapé).
-            float& a_ii = maximo_abs;
+            double& a_ii = maximo_abs;
             for (auto iter_lista = ++filas_a_ser_restadas.begin(); iter_lista != filas_a_ser_restadas.end(); ++iter_lista) {
-                float& a_ji = iter_lista->second;
+                double& a_ji = iter_lista->second;
                 copy.resta_de_filas(iter_lista->first, a_ji / a_ii, i);
             }
         }
     }
     for(i = 0; i < copy.cantFilas(); i++){
         j = copy.cantFilas()-1-i;
-        float a_jj = copy.at(j,j);
+        double a_jj = copy.at(j,j);
         if(a_jj == 0 && bb[j] != 0){
             status = -1; //El sistema es incompatible.
             break;
@@ -292,7 +292,7 @@ pair<vector<float>,short> VectorMapMatrix::EGPP(vector<float> bb) {
     return make_pair(res,status);
 }
 
-void VectorMapMatrix::resta_de_filas(uint fila_a_modificar, float escalar, uint fila_para_restar){
+void VectorMapMatrix::resta_de_filas(uint fila_a_modificar, double escalar, uint fila_para_restar){
     if(escalar == 0) return;
     else {
         auto it1 = m[fila_a_modificar].begin();
@@ -300,7 +300,7 @@ void VectorMapMatrix::resta_de_filas(uint fila_a_modificar, float escalar, uint 
         auto it2 = m[fila_para_restar].begin();
         auto fin2 = m[fila_para_restar].end();
         while (it2 != fin2) {  //Mientras no haya acabado la fila para restar:
-            float resultado_de_la_resta = -escalar * (it2->second); //empiezo restando lo que hay que restar
+            double resultado_de_la_resta = -escalar * (it2->second); //empiezo restando lo que hay que restar
             while (it1 != fin1 && it1->first < it2->first) ++it1;
             if (it1->first == it2->first) resultado_de_la_resta += it1->second; //si hay algo distinto de 0 en la coordenada a modificar, lo sumo
             if (resultado_de_la_resta != 0) m[fila_a_modificar][it2->first] = resultado_de_la_resta;    //si el resultado de la resta no es 0, lo defino en el diccionario/fila
@@ -312,7 +312,7 @@ void VectorMapMatrix::resta_de_filas(uint fila_a_modificar, float escalar, uint 
  * Funciones para mostrar la matriz
  */
 
-string convertirAString(float num) {
+string convertirAString(double num) {
     stringstream ss (stringstream::in | stringstream::out);
     ss << num;
     string cadena = ss.str();
@@ -332,7 +332,7 @@ int cantidadDeDigitosMaxima(VectorMapMatrix &M) {
     return maximo;
 }
 
-string agregarEspacios(float valor, int cantidadMaxima) {
+string agregarEspacios(double valor, int cantidadMaxima) {
     string ret = convertirAString(valor);
     cantidadMaxima = cantidadMaxima - ret.length();
     for (int i=0; i < cantidadMaxima+1; i++) {
